Fix swagger and remove dispatch_request.
